+++ server/events/command_runner.go
@@ -84,12 +84,17 @@ type DefaultCommandRunner struct {
 	// this in our error message back to the user on a forked PR so they know
 	// how to enable this functionality.
 	AllowForkPRsFlag string
+	// HidePrevPlanComments will hide previous plan comments to declutter PRs.
+	HidePrevPlanComments bool
 	// SilenceForkPRErrors controls whether to comment on Fork PRs when AllowForkPRs = False
 	SilenceForkPRErrors bool
 	// SilenceForkPRErrorsFlag is the name of the flag that controls fork PR's. We use
 	// this in our error message back to the user on a forked PR so they know
 	// how to disable error comment
 	SilenceForkPRErrorsFlag string
+	// SilenceVCSStatusNoPlans is whether autoplan should set commit status if no plans
+	// are found
+	SilenceVCSStatusNoPlans bool
 	ProjectCommandBuilder   ProjectCommandBuilder
 	ProjectCommandRunner    ProjectCommandRunner
 	// GlobalAutomerge is true if we should automatically merge pull requests if all
@@ -115,10 +120,6 @@ func (c *DefaultCommandRunner) RunAutoplanCommand(baseRepo models.Repo, headRepo
 		return
 	}
 
-	if err := c.CommitStatusUpdater.UpdateCombined(ctx.BaseRepo, ctx.Pull, models.PendingCommitStatus, models.PlanCommand); err != nil {
-		ctx.Log.Warn("unable to update commit status: %s", err)
-	}
-
 	projectCmds, err := c.ProjectCommandBuilder.BuildAutoplanCommands(ctx)
 	if err != nil {
 		if statusErr := c.CommitStatusUpdater.UpdateCombined(ctx.BaseRepo, ctx.Pull, models.FailedCommitStatus, models.PlanCommand); statusErr != nil {
@@ -130,13 +131,21 @@ func (c *DefaultCommandRunner) RunAutoplanCommand(baseRepo models.Repo, headRepo
 	}
 	if len(projectCmds) == 0 {
 		log.Info("determined there was no project to run plan in")
-		// If there were no projects modified, we set a successful commit status
-		// with 0/0 projects planned successfully because we've already set an
-		// in-progress status and we don't want that to be "in progress" forever.
-		if err := c.CommitStatusUpdater.UpdateCombinedCount(baseRepo, pull, models.SuccessCommitStatus, models.PlanCommand, 0, 0); err != nil {
-			ctx.Log.Warn("unable to update commit status: %s", err)
+		if !c.SilenceVCSStatusNoPlans {
+			// If there were no projects modified, we set a successful commit status
+			// with 0/0 projects planned successfully because some users require
+			// the Atlantis status to be passing for all pull requests.
+			ctx.Log.Debug("setting VCS status to success with no projects found")
+			if err := c.CommitStatusUpdater.UpdateCombinedCount(baseRepo, pull, models.SuccessCommitStatus, models.PlanCommand, 0, 0); err != nil {
+				ctx.Log.Warn("unable to update commit status: %s", err)
+			}
 		}
 		return
+	}
+
+	// At this point we are sure Atlantis has work to do, so set commit status to pending
+	if err := c.CommitStatusUpdater.UpdateCombined(ctx.BaseRepo, ctx.Pull, models.PendingCommitStatus, models.PlanCommand); err != nil {
+		ctx.Log.Warn("unable to update commit status: %s", err)
 	}
 
 	// Run our plan commands in parallel if enabled
@@ -194,6 +203,7 @@ func (c *DefaultCommandRunner) RunCommentCommand(baseRepo models.Repo, maybeHead
 	case models.BitbucketCloud, models.BitbucketServer:
 		if maybePull == nil {
 			err = errors.New("pull request should not be nil–this is a bug")
+			break
 		}
 		pull = *maybePull
 	case models.AzureDevops:
@@ -474,6 +484,15 @@ func (c *DefaultCommandRunner) updatePull(ctx *CommandContext, command PullComma
 		ctx.Log.Err(res.Error.Error())
 	} else if res.Failure != "" {
 		ctx.Log.Warn(res.Failure)
+	}
+
+	// HidePrevPlanComments will hide old comments left from previous plan runs to reduce
+	// clutter in a pull/merge request. This will not delete the comment, since the
+	// comment trail may be useful in auditing or backtracing problems.
+	if c.HidePrevPlanComments {
+		if err := c.VCSClient.HidePrevPlanComments(ctx.BaseRepo, ctx.Pull.Num); err != nil {
+			ctx.Log.Err("unable to hide old comments: %s", err)
+		}
 	}
 
 	comment := c.MarkdownRenderer.Render(res, command.CommandName(), ctx.Log.History.String(), command.IsVerbose(), ctx.BaseRepo.VCSHost.Type)