+++ server/events/vcs/client.go
@@ -25,6 +25,7 @@ type Client interface {
 	// relative to the repo root, e.g. parent/child/file.txt.
 	GetModifiedFiles(repo models.Repo, pull models.PullRequest) ([]string, error)
 	CreateComment(repo models.Repo, pullNum int, comment string) error
+	HidePrevPlanComments(repo models.Repo, pullNum int) error
 	PullIsApproved(repo models.Repo, pull models.PullRequest) (bool, error)
 	PullIsMergeable(repo models.Repo, pull models.PullRequest) (bool, error)
 	// UpdateStatus updates the commit status to state for pull. src is the
@@ -36,4 +37,5 @@ type Client interface {
 	// about this status.
 	UpdateStatus(repo models.Repo, pull models.PullRequest, state models.CommitStatus, src string, description string, url string) error
 	MergePull(pull models.PullRequest) error
+	MarkdownPullLink(pull models.PullRequest) (string, error)
 }