+++ server/events/vcs/not_configured_vcs_client.go
@@ -32,6 +32,9 @@ func (a *NotConfiguredVCSClient) GetModifiedFiles(repo models.Repo, pull models.
 func (a *NotConfiguredVCSClient) CreateComment(repo models.Repo, pullNum int, comment string) error {
 	return a.err()
 }
+func (a *NotConfiguredVCSClient) HidePrevPlanComments(repo models.Repo, pullNum int) error {
+	return nil
+}
 func (a *NotConfiguredVCSClient) PullIsApproved(repo models.Repo, pull models.PullRequest) (bool, error) {
 	return false, a.err()
 }
@@ -43,6 +46,9 @@ func (a *NotConfiguredVCSClient) UpdateStatus(repo models.Repo, pull models.Pull
 }
 func (a *NotConfiguredVCSClient) MergePull(pull models.PullRequest) error {
 	return a.err()
+}
+func (a *NotConfiguredVCSClient) MarkdownPullLink(pull models.PullRequest) (string, error) {
+	return "", a.err()
 }
 func (a *NotConfiguredVCSClient) err() error {
 	return fmt.Errorf("atlantis was not configured to support repos from %s", a.Host.String())