+++ server/events/vcs/proxy.go
@@ -60,6 +60,10 @@ func (d *ClientProxy) CreateComment(repo models.Repo, pullNum int, comment strin
 	return d.clients[repo.VCSHost.Type].CreateComment(repo, pullNum, comment)
 }
 
+func (d *ClientProxy) HidePrevPlanComments(repo models.Repo, pullNum int) error {
+	return d.clients[repo.VCSHost.Type].HidePrevPlanComments(repo, pullNum)
+}
+
 func (d *ClientProxy) PullIsApproved(repo models.Repo, pull models.PullRequest) (bool, error) {
 	return d.clients[repo.VCSHost.Type].PullIsApproved(repo, pull)
 }
@@ -74,4 +78,8 @@ func (d *ClientProxy) UpdateStatus(repo models.Repo, pull models.PullRequest, st
 
 func (d *ClientProxy) MergePull(pull models.PullRequest) error {
 	return d.clients[pull.BaseRepo.VCSHost.Type].MergePull(pull)
+}
+
+func (d *ClientProxy) MarkdownPullLink(pull models.PullRequest) (string, error) {
+	return d.clients[pull.BaseRepo.VCSHost.Type].MarkdownPullLink(pull)
 }