+++ server/server.go
@@ -231,7 +231,9 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 	markdownRenderer := &events.MarkdownRenderer{
 		GitlabSupportsCommonMark: gitlabClient.SupportsCommonMark(),
 		DisableApplyAll:          userConfig.DisableApplyAll,
+		DisableMarkdownFolding:   userConfig.DisableMarkdownFolding,
 	}
+
 	boltdb, err := db.New(userConfig.DataDir)
 	if err != nil {
 		return nil, err
@@ -243,7 +245,8 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 		CheckoutMerge: userConfig.CheckoutStrategy == "merge",
 	}
 	projectLocker := &events.DefaultProjectLocker{
-		Locker: lockingClient,
+		Locker:    lockingClient,
+		VCSClient: vcsClient,
 	}
 	parsedURL, err := ParseAtlantisURL(userConfig.AtlantisURL)
 	if err != nil {
@@ -314,8 +317,10 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 		Logger:                   logger,
 		AllowForkPRs:             userConfig.AllowForkPRs,
 		AllowForkPRsFlag:         config.AllowForkPRsFlag,
+		HidePrevPlanComments:     userConfig.HidePrevPlanComments,
 		SilenceForkPRErrors:      userConfig.SilenceForkPRErrors,
 		SilenceForkPRErrorsFlag:  config.SilenceForkPRErrorsFlag,
+		SilenceVCSStatusNoPlans:  userConfig.SilenceVCSStatusNoPlans,
 		DisableApplyAll:          userConfig.DisableApplyAll,
 		ParallelPlansPoolSize:    userConfig.ParallelPlansPoolSize,
 		ProjectCommandBuilder: &events.DefaultProjectCommandBuilder{