+++ server/user_config.go
@@ -1,6 +1,8 @@
 package server
 
-import "github.com/runatlantis/atlantis/server/logging"
+import (
+	"github.com/runatlantis/atlantis/server/logging"
+)
 
 // UserConfig holds config values passed in by the user.
 // The mapstructure tags correspond to flags in cmd/server.go and are used when
@@ -21,6 +23,7 @@ type UserConfig struct {
 	CheckoutStrategy           string `mapstructure:"checkout-strategy"`
 	DataDir                    string `mapstructure:"data-dir"`
 	DisableApplyAll            bool   `mapstructure:"disable-apply-all"`
+	DisableMarkdownFolding     bool   `mapstructure:"disable-markdown-folding"`
 	GithubHostname             string `mapstructure:"gh-hostname"`
 	GithubToken                string `mapstructure:"gh-token"`
 	GithubUser                 string `mapstructure:"gh-user"`
@@ -29,6 +32,7 @@ type UserConfig struct {
 	GitlabToken                string `mapstructure:"gitlab-token"`
 	GitlabUser                 string `mapstructure:"gitlab-user"`
 	GitlabWebhookSecret        string `mapstructure:"gitlab-webhook-secret"`
+	HidePrevPlanComments       bool   `mapstructure:"hide-prev-plan-comments"`
 	LogLevel                   string `mapstructure:"log-level"`
 	ParallelPlansPoolSize      int    `mapstructure:"parallel-plans-pool-size"`
 	Port                       int    `mapstructure:"port"`
@@ -40,19 +44,22 @@ type UserConfig struct {
 	RequireApproval bool `mapstructure:"require-approval"`
 	// RequireMergeable is whether to require pull requests to be mergeable before
 	// allowing terraform apply's to run.
-	RequireMergeable       bool            `mapstructure:"require-mergeable"`
-	SilenceForkPRErrors    bool            `mapstructure:"silence-fork-pr-errors"`
-	SilenceWhitelistErrors bool            `mapstructure:"silence-whitelist-errors"`
-	SlackToken             string          `mapstructure:"slack-token"`
-	SSLCertFile            string          `mapstructure:"ssl-cert-file"`
-	SSLKeyFile             string          `mapstructure:"ssl-key-file"`
-	TFDownloadURL          string          `mapstructure:"tf-download-url"`
-	TFEHostname            string          `mapstructure:"tfe-hostname"`
-	TFEToken               string          `mapstructure:"tfe-token"`
-	VCSStatusName          string          `mapstructure:"vcs-status-name"`
-	DefaultTFVersion       string          `mapstructure:"default-tf-version"`
-	Webhooks               []WebhookConfig `mapstructure:"webhooks"`
-	WriteGitCreds          bool            `mapstructure:"write-git-creds"`
+	RequireMergeable    bool `mapstructure:"require-mergeable"`
+	SilenceForkPRErrors bool `mapstructure:"silence-fork-pr-errors"`
+	// SilenceVCSStatusNoPlans is whether autoplan should set commit status if no plans
+	// are found.
+	SilenceVCSStatusNoPlans bool            `mapstructure:"silence-vcs-status-no-plans"`
+	SilenceWhitelistErrors  bool            `mapstructure:"silence-whitelist-errors"`
+	SlackToken              string          `mapstructure:"slack-token"`
+	SSLCertFile             string          `mapstructure:"ssl-cert-file"`
+	SSLKeyFile              string          `mapstructure:"ssl-key-file"`
+	TFDownloadURL           string          `mapstructure:"tf-download-url"`
+	TFEHostname             string          `mapstructure:"tfe-hostname"`
+	TFEToken                string          `mapstructure:"tfe-token"`
+	VCSStatusName           string          `mapstructure:"vcs-status-name"`
+	DefaultTFVersion        string          `mapstructure:"default-tf-version"`
+	Webhooks                []WebhookConfig `mapstructure:"webhooks"`
+	WriteGitCreds           bool            `mapstructure:"write-git-creds"`
 }
 
 // ToLogLevel returns the LogLevel object corresponding to the user-passed